refactor: move step 4 of submit to own file (#654)
**Context:**



**Changes In This Pull Request:**



**Test Plan:**
:
